// src/models/schema/FieldType.ts
export enum FieldType {
	object = "object",
	string = "string",
	integer = "integer",
	float = "float",
	boolean = "boolean",
	array = "array",
	enum = "enum",
	tags = "tags",
	const = "const",
	json = "json",
	date = "date",
	datetime = "datetime",
	longtext = "longtext",
	richtext = "richtext",
	email = "email",
	uri = "uri",
	hostname = "hostname",
	color = "color",
	location = "location",
	code = "code",
	image = "image",
}